+++ vonage_utils/src/vonage_utils/__init__.py
@@ -1,4 +1,5 @@
 from .errors import VonageError
+from .types.phone_number import PhoneNumber
 from .utils import format_phone_number, remove_none_values
 
-__all__ = ['VonageError', 'format_phone_number', 'remove_none_values']
+__all__ = ['VonageError', 'format_phone_number', 'remove_none_values', PhoneNumber]