+++ vonage/CHANGES.md
@@ -1,5 +1,6 @@
 # 3.99.0a4
 - Add support for the [Vonage Verify API](https://developer.vonage.com/en/api/verify).
+- Add `last_request` and `last_response` properties to the HTTP Client.
 
 # 3.99.0a3
 - Add support for the [Vonage Users API](https://developer.vonage.com/en/api/application.v2#User).