+++ vonage_utils/CHANGES.md
@@ -1,2 +1,5 @@
+# 1.0.1
+- Add `PhoneNumber` type
+
 # 1.0.0
 - Initial upload